chore(deps): refactor -- rename utils-* to *_ops
