+++ rpxy-lib/src/message_handler/header_ops/forwarding.rs
@@ -1,20 +1,17 @@
-use super::canonical_address::ToCanonical;
-use crate::{
-  backend::{UpstreamCandidates, UpstreamOption},
-  log::*,
-};
 use anyhow::{Result, anyhow};
 use http::{HeaderMap, HeaderName, HeaderValue, Uri, header, header::AsHeaderName};
 use ipnet::IpNet;
 use std::{
-  borrow::Cow,
   net::{IpAddr, SocketAddr},
   str::FromStr,
 };
 
-#[cfg(feature = "sticky-cookie")]
-use crate::backend::{LoadBalanceContext, StickyCookie, StickyCookieValue};
-// use crate::backend::{UpstreamGroup, UpstreamOption};
+use crate::log::*;
+
+use super::{
+  super::canonical_address::ToCanonical,
+  common::{add_header_entry_overwrite_if_exist, host_from_uri_or_host_header, make_cookie_single_line},
+};
 
 const X_FORWARDED_FOR: &str = "x-forwarded-for";
 const X_FORWARDED_PROTO: &str = "x-forwarded-proto";
@@ -22,177 +19,6 @@ const X_FORWARDED_PORT: &str = "x-forwarded-port";
 const X_FORWARDED_SSL: &str = "x-forwarded-ssl";
 const X_ORIGINAL_URI: &str = "x-original-uri";
 const X_REAL_IP: &str = "x-real-ip";
-
-/* --------------------------------------------------------------------------------------------------------- */
-/* Request / response header helpers                                                                       */
-/* --------------------------------------------------------------------------------------------------------- */
-
-#[cfg(feature = "sticky-cookie")]
-/// Take sticky cookie header value from request header,
-/// and returns LoadBalanceContext to be forwarded to LB if exist and if needed.
-/// Removing sticky cookie is needed and it must not be passed to the upstream.
-pub(super) fn takeout_sticky_cookie_lb_context(
-  headers: &mut HeaderMap,
-  expected_cookie_name: &str,
-) -> Result<Option<LoadBalanceContext>> {
-  let mut headers_clone = headers.clone();
-
-  match headers_clone.entry(header::COOKIE) {
-    header::Entry::Vacant(_) => Ok(None),
-    header::Entry::Occupied(entry) => {
-      let cookies_iter = entry
-        .iter()
-        .flat_map(|v| v.to_str().unwrap_or("").split(';').map(|v| v.trim()));
-      let (sticky_cookies, without_sticky_cookies): (Vec<_>, Vec<_>) =
-        cookies_iter.into_iter().partition(|v| v.starts_with(expected_cookie_name));
-      if sticky_cookies.is_empty() {
-        return Ok(None);
-      }
-      anyhow::ensure!(sticky_cookies.len() == 1, "Invalid cookie: Multiple sticky cookie values");
-
-      let cookies_passed_to_upstream = without_sticky_cookies.join("; ");
-      let cookie_passed_to_lb = sticky_cookies.first().unwrap();
-      headers.remove(header::COOKIE);
-      headers.insert(header::COOKIE, cookies_passed_to_upstream.parse()?);
-
-      let sticky_cookie = StickyCookie {
-        value: StickyCookieValue::try_from(cookie_passed_to_lb, expected_cookie_name)?,
-        info: None,
-      };
-      Ok(Some(LoadBalanceContext { sticky_cookie }))
-    }
-  }
-}
-
-#[cfg(feature = "sticky-cookie")]
-/// Set-Cookie if LB Sticky is enabled and if cookie is newly created/updated.
-/// Set-Cookie response header could be in multiple lines.
-/// https://developer.mozilla.org/ja/docs/Web/HTTP/Headers/Set-Cookie
-pub(super) fn set_sticky_cookie_lb_context(headers: &mut HeaderMap, context_from_lb: &LoadBalanceContext) -> Result<()> {
-  let sticky_cookie_string: String = context_from_lb.sticky_cookie.clone().try_into()?;
-  let new_header_val: HeaderValue = sticky_cookie_string.parse()?;
-  let expected_cookie_name = &context_from_lb.sticky_cookie.value.name;
-  match headers.entry(header::SET_COOKIE) {
-    header::Entry::Vacant(entry) => {
-      entry.insert(new_header_val);
-    }
-    header::Entry::Occupied(mut entry) => {
-      let mut flag = false;
-      for e in entry.iter_mut() {
-        if e.to_str().unwrap_or("").starts_with(expected_cookie_name) {
-          *e = new_header_val.clone();
-          flag = true;
-        }
-      }
-      if !flag {
-        entry.append(new_header_val);
-      }
-    }
-  };
-  Ok(())
-}
-
-/// overwrite HOST value with upstream hostname (like 192.168.xx.x seen from rpxy)
-fn override_host_header(headers: &mut HeaderMap, upstream_base_uri: &Uri) -> Result<()> {
-  let mut upstream_host = upstream_base_uri
-    .host()
-    .ok_or_else(|| anyhow!("No hostname is given"))?
-    .to_string();
-  // add port if it is not default
-  if let Some(port) = upstream_base_uri.port_u16() {
-    upstream_host = format!("{}:{}", upstream_host, port);
-  }
-
-  // overwrite host header, this removes all the HOST header values
-  headers.insert(header::HOST, HeaderValue::from_str(&upstream_host)?);
-  Ok(())
-}
-
-/// Apply options to request header, which are specified in the configuration
-/// This function is called after almost all other headers has been set and updated.
-pub(super) fn apply_upstream_options_to_header(
-  headers: &mut HeaderMap,
-  upstream_base_uri: &Uri,
-  // _client_addr: &SocketAddr,
-  upstream: &UpstreamCandidates,
-  trusted_forwarded_proxies: &[IpNet],
-) -> Result<()> {
-  for opt in upstream.options.iter() {
-    match opt {
-      UpstreamOption::SetUpstreamHost => {
-        // prioritize KeepOriginalHost
-        if !upstream.options.contains(&UpstreamOption::KeepOriginalHost) {
-          // overwrite host header, this removes all the HOST header values
-          override_host_header(headers, upstream_base_uri)?;
-        }
-      }
-      UpstreamOption::UpgradeInsecureRequests => {
-        // add upgrade-insecure-requests in request header if not exist
-        headers
-          .entry(header::UPGRADE_INSECURE_REQUESTS)
-          .or_insert(HeaderValue::from_bytes(b"1").unwrap());
-      }
-      UpstreamOption::ForwardedHeader => {
-        // This is called after X-Forwarded-For is added to generate RFC 7239 Forwarded header from it.
-        // If Forwarded already exists, it has already been normalized by add_forwarding_header().
-        if !headers.contains_key(header::FORWARDED) {
-          let Some(forwarding_chain) = extract_forwarding_chain_from_headers(headers)? else {
-            warn!("Failed to generate Forwarded header: no X-Forwarded-For information found in headers");
-            continue;
-          };
-          let normalized_chain = reduce_trusted_proxy_chain(forwarding_chain, trusted_forwarded_proxies);
-          match generate_forwarded_header(&normalized_chain) {
-            Ok(forwarded_value) => {
-              add_header_entry_overwrite_if_exist(headers, header::FORWARDED.as_str(), forwarded_value)?;
-            }
-            Err(e) => {
-              // Log warning but don't fail the request if Forwarded generation fails
-              warn!("Failed to generate Forwarded header: {}", e);
-            }
-          }
-        }
-      }
-      _ => (),
-    }
-  }
-
-  Ok(())
-}
-
-/// Overwrite header entry if exist
-pub(super) fn add_header_entry_overwrite_if_exist(
-  headers: &mut HeaderMap,
-  key: impl Into<Cow<'static, str>>,
-  value: impl Into<Cow<'static, str>>,
-) -> Result<()> {
-  match headers.entry(HeaderName::from_bytes(key.into().as_bytes())?) {
-    header::Entry::Vacant(entry) => {
-      entry.insert(value.into().parse::<HeaderValue>()?);
-    }
-    header::Entry::Occupied(mut entry) => {
-      entry.insert(HeaderValue::from_bytes(value.into().as_bytes())?);
-    }
-  }
-
-  Ok(())
-}
-
-/// Align cookie values in single line
-/// Sometimes violates [RFC6265](https://www.rfc-editor.org/rfc/rfc6265#section-5.4) (for http/1.1).
-/// This is allowed in RFC7540 (for http/2) as mentioned [here](https://stackoverflow.com/questions/4843556/in-http-specification-what-is-the-string-that-separates-cookies).
-pub(super) fn make_cookie_single_line(headers: &mut HeaderMap) -> Result<()> {
-  let cookies = headers
-    .iter()
-    .filter(|(k, _)| **k == header::COOKIE)
-    .map(|(_, v)| v.to_str().unwrap_or(""))
-    .collect::<Vec<_>>()
-    .join("; ");
-  if !cookies.is_empty() {
-    headers.remove(header::COOKIE);
-    headers.insert(header::COOKIE, HeaderValue::from_bytes(cookies.as_bytes())?);
-  }
-  Ok(())
-}
 
 /* --------------------------------------------------------------------------------------------------------- */
 /* Forwarding header normalization model                                                                   */
@@ -211,23 +37,27 @@ pub(super) fn make_cookie_single_line(headers: &mut HeaderMap) -> Result<()> {
 
 /// An entry in Forwarded header with only the parameters relevant for forwarding chain normalization and consistency check.
 #[derive(Clone, Debug, PartialEq, Eq)]
-enum ForwardedNode {
-  Ip(IpAddr, Option<u16>),
-  Unknown(Option<String>),
-  Obfuscated(String, Option<String>),
+struct ForwardedNode {
+  ip: Option<IpAddr>,
+  port: Option<String>,
+  raw: ForwardedNodeRaw,
+}
+
+#[derive(Clone, Debug, PartialEq, Eq)]
+enum ForwardedNodeRaw {
+  Ip,
+  Unknown,
+  Obfuscated(String),
 }
 
 impl ForwardedNode {
   fn ip_addr(&self) -> Option<IpAddr> {
-    match self {
-      Self::Ip(ip, _) => Some(*ip),
-      Self::Unknown(_) | Self::Obfuscated(_, _) => None,
-    }
+    self.ip
   }
 }
 
 #[derive(Clone, Debug, PartialEq, Eq)]
-struct ForwardedEntry {
+pub(super) struct ForwardedEntry {
   /// for=:
   /// For Forwarded Header: this is optional in the RFC, but we require it for each entry to be able to build the forwarding chain. If missing, the entry is considered invalid.
   for_node: ForwardedNode,
@@ -245,7 +75,7 @@ struct ForwardedEntry {
 /// If the immediate peer is in trusted_forwarded_proxies,
 /// it will recursively trust the incoming forwarding headers and append the immediate peer IP to the end of the chain.
 /// Otherwise, it will ignore incoming forwarding headers and start a new chain with the immediate peer IP.
-pub(super) fn add_forwarding_header(
+pub(in crate::message_handler) fn add_forwarding_header(
   headers: &mut HeaderMap,
   client_addr: &SocketAddr,
   listen_addr: &SocketAddr,
@@ -348,7 +178,7 @@ fn normalize_forwarding_chain(
 /// Extract forwarding information chain from Forwarded or X-Forwarded-For/Proto headers.
 /// Returns None if neither header is present.
 /// If both are present, prefer Forwarded only when it is consistent with the auxiliary X-Forwarded-* view.
-fn extract_forwarding_chain_from_headers(headers: &HeaderMap) -> Result<Option<Vec<ForwardedEntry>>> {
+pub(super) fn extract_forwarding_chain_from_headers(headers: &HeaderMap) -> Result<Option<Vec<ForwardedEntry>>> {
   let xff_chain = if headers.contains_key(X_FORWARDED_FOR) {
     Some(parse_x_forwarded_for_header(headers)?)
   } else {
@@ -384,7 +214,10 @@ fn extract_forwarding_chain_from_headers(headers: &HeaderMap) -> Result<Option<V
 
 /// Reduce the forwarding chain by removing trusted proxy hops from the end.
 /// We assume the immediate peer is always appended as the last hop.
-fn reduce_trusted_proxy_chain(mut chain: Vec<ForwardedEntry>, trusted_forwarded_proxies: &[IpNet]) -> Vec<ForwardedEntry> {
+pub(super) fn reduce_trusted_proxy_chain(
+  mut chain: Vec<ForwardedEntry>,
+  trusted_forwarded_proxies: &[IpNet],
+) -> Vec<ForwardedEntry> {
   let mut idx = chain.len();
   while idx > 0 && entry_is_trusted_proxy(&chain[idx - 1], trusted_forwarded_proxies) {
     idx -= 1;
@@ -412,7 +245,11 @@ fn parse_x_forwarded_for_header(headers: &HeaderMap) -> Result<Vec<ForwardedEntr
   for (idx, ip) in ips.into_iter().enumerate() {
     let for_ip = parse_forwarded_ip_token(ip)?;
     chain.push(ForwardedEntry {
-      for_node: ForwardedNode::Ip(for_ip, None),
+      for_node: ForwardedNode {
+        ip: Some(for_ip),
+        port: None,
+        raw: ForwardedNodeRaw::Ip,
+      },
       proto: if idx == last_idx { xf_proto.clone() } else { None },
       host: None,
       by: None,
@@ -603,8 +440,8 @@ fn unquote_http_value(value: &str) -> Result<String> {
 /// Parse a token in Forwarded header that is supposed to represent a client IP, which can be an IP address, "unknown", or obfuscated identifier.
 fn parse_forwarded_ip_token(token: &str) -> Result<IpAddr> {
   match parse_forwarded_node(token)? {
-    ForwardedNode::Ip(ip, _) => Ok(ip),
-    ForwardedNode::Unknown(_) | ForwardedNode::Obfuscated(_, _) => Err(anyhow!("forwarded node `{token}` is not an IP address")),
+    ForwardedNode { ip: Some(ip), .. } => Ok(ip),
+    ForwardedNode { ip: None, .. } => Err(anyhow!("forwarded node `{token}` is not an IP address")),
   }
 }
 
@@ -612,35 +449,64 @@ fn parse_forwarded_ip_token(token: &str) -> Result<IpAddr> {
 fn parse_forwarded_node(token: &str) -> Result<ForwardedNode> {
   let trimmed = unquote_http_value(token)?;
   if trimmed.eq_ignore_ascii_case("unknown") {
-    return Ok(ForwardedNode::Unknown(None));
+    return Ok(ForwardedNode {
+      ip: None,
+      port: None,
+      raw: ForwardedNodeRaw::Unknown,
+    });
   }
   if let Some((node, port)) = trimmed.rsplit_once(':') {
     if let Some(inner) = node.strip_prefix('[').and_then(|s| s.strip_suffix(']')) {
       let ip = canonicalize_ip(IpAddr::from_str(inner).map_err(|e| anyhow!("invalid forwarded address `{token}`: {e}"))?);
-      return Ok(ForwardedNode::Ip(ip, Some(parse_forwarded_port(port, token)?)));
+      return Ok(ForwardedNode {
+        ip: Some(ip),
+        port: Some(parse_forwarded_port(port, token)?.to_string()),
+        raw: ForwardedNodeRaw::Ip,
+      });
     }
     if node.eq_ignore_ascii_case("unknown") {
-      return Ok(ForwardedNode::Unknown(Some(port.to_string())));
+      return Ok(ForwardedNode {
+        ip: None,
+        port: Some(port.to_string()),
+        raw: ForwardedNodeRaw::Unknown,
+      });
     }
     if node.starts_with('_') {
-      return Ok(ForwardedNode::Obfuscated(node.to_string(), Some(port.to_string())));
+      return Ok(ForwardedNode {
+        ip: None,
+        port: Some(port.to_string()),
+        raw: ForwardedNodeRaw::Obfuscated(node.to_string()),
+      });
     }
     if let Ok(ip) = IpAddr::from_str(node) {
-      return Ok(ForwardedNode::Ip(
-        canonicalize_ip(ip),
-        Some(parse_forwarded_port(port, token)?),
-      ));
+      return Ok(ForwardedNode {
+        ip: Some(canonicalize_ip(ip)),
+        port: Some(parse_forwarded_port(port, token)?.to_string()),
+        raw: ForwardedNodeRaw::Ip,
+      });
     }
   }
 
   if let Some(inner) = trimmed.strip_prefix('[').and_then(|s| s.strip_suffix(']')) {
     let ip = IpAddr::from_str(inner).map_err(|e| anyhow!("invalid forwarded address `{token}`: {e}"))?;
-    Ok(ForwardedNode::Ip(canonicalize_ip(ip), None))
+    Ok(ForwardedNode {
+      ip: Some(canonicalize_ip(ip)),
+      port: None,
+      raw: ForwardedNodeRaw::Ip,
+    })
   } else if trimmed.starts_with('_') {
-    Ok(ForwardedNode::Obfuscated(trimmed, None))
+    Ok(ForwardedNode {
+      ip: None,
+      port: None,
+      raw: ForwardedNodeRaw::Obfuscated(trimmed),
+    })
   } else {
     let ip = IpAddr::from_str(&trimmed).map_err(|e| anyhow!("invalid forwarded address `{token}`: {e}"))?;
-    Ok(ForwardedNode::Ip(canonicalize_ip(ip), None))
+    Ok(ForwardedNode {
+      ip: Some(canonicalize_ip(ip)),
+      port: None,
+      raw: ForwardedNodeRaw::Ip,
+    })
   }
 }
 
@@ -663,7 +529,11 @@ fn canonicalize_ip(ip: IpAddr) -> IpAddr {
 /// Build a ForwardedEntry for the immediate peer, which represents this proxy's authoritative view of the client connection. This is used as the basis for forwarding chain normalization and generation.
 fn build_peer_forwarded_entry(headers: &HeaderMap, peer_ip: IpAddr, tls: bool, original_uri: &Uri) -> ForwardedEntry {
   ForwardedEntry {
-    for_node: ForwardedNode::Ip(peer_ip, None),
+    for_node: ForwardedNode {
+      ip: Some(peer_ip),
+      port: None,
+      raw: ForwardedNodeRaw::Ip,
+    },
     proto: if tls { Some("https".into()) } else { Some("http".into()) },
     host: host_from_uri_or_host_header(original_uri, headers.get(header::HOST).cloned()).ok(),
     by: None,
@@ -705,7 +575,7 @@ fn overwrite_header_with_csv(headers: &mut HeaderMap, key: &str, values: &[Strin
 /// This function assumes that the X-Forwarded-For header is present and well-formed.
 /// Earlier hops are emitted as `for=` only, and the last hop carries this proxy's
 /// authoritative `proto` and `host` parameters.
-fn generate_forwarded_header(normalized_forwarding_chain: &[ForwardedEntry]) -> Result<String> {
+pub(super) fn generate_forwarded_header(normalized_forwarding_chain: &[ForwardedEntry]) -> Result<String> {
   if normalized_forwarding_chain.is_empty() {
     return Err(anyhow!("No forwarding chain found for Forwarded generation"));
   }
@@ -733,97 +603,19 @@ fn generate_forwarded_header(normalized_forwarding_chain: &[ForwardedEntry]) -> 
 
 /// Format the for= value in Forwarded header according to RFC 7239, which may require quoting and bracketing for IPv6 addresses.
 fn format_forwarded_node(node: &ForwardedNode) -> Result<String> {
-  match node {
-    ForwardedNode::Ip(IpAddr::V4(v4), None) => Ok(v4.to_string()),
-    ForwardedNode::Ip(IpAddr::V4(v4), Some(port)) => Ok(format!("\"{}:{}\"", v4, port)),
-    ForwardedNode::Ip(IpAddr::V6(v6), None) => Ok(format!("\"[{}]\"", v6)),
-    ForwardedNode::Ip(IpAddr::V6(v6), Some(port)) => Ok(format!("\"[{}]:{}\"", v6, port)),
-    ForwardedNode::Unknown(None) => Ok("unknown".to_string()),
-    ForwardedNode::Unknown(Some(port)) => Ok(format!("\"unknown:{}\"", port)),
-    ForwardedNode::Obfuscated(node, None) => Ok(node.clone()),
-    ForwardedNode::Obfuscated(node, Some(port)) => Ok(format!("\"{}:{}\"", node, port)),
+  match (&node.raw, node.ip, node.port.as_deref()) {
+    (ForwardedNodeRaw::Ip, Some(IpAddr::V4(v4)), None) => Ok(v4.to_string()),
+    (ForwardedNodeRaw::Ip, Some(IpAddr::V4(v4)), Some(port)) => Ok(format!("\"{}:{}\"", v4, port)),
+    (ForwardedNodeRaw::Ip, Some(IpAddr::V6(v6)), None) => Ok(format!("\"[{}]\"", v6)),
+    (ForwardedNodeRaw::Ip, Some(IpAddr::V6(v6)), Some(port)) => Ok(format!("\"[{}]:{}\"", v6, port)),
+    (ForwardedNodeRaw::Unknown, None, None) => Ok("unknown".to_string()),
+    (ForwardedNodeRaw::Unknown, None, Some(port)) => Ok(format!("\"unknown:{}\"", port)),
+    (ForwardedNodeRaw::Obfuscated(node), None, None) => Ok(node.clone()),
+    (ForwardedNodeRaw::Obfuscated(node), None, Some(port)) => Ok(format!("\"{}:{}\"", node, port)),
+    _ => Err(anyhow!("invalid forwarded node state")),
   }
 }
 
-/* --------------------------------------------------------------------------------------------------------- */
-/* Generic URI / connection header helpers                                                                 */
-/* --------------------------------------------------------------------------------------------------------- */
-
-/// Extract host from URI
-pub(super) fn host_from_uri_or_host_header(uri: &Uri, host_header_value: Option<header::HeaderValue>) -> Result<String> {
-  // Prioritize uri host over host header
-  let uri_host = uri.host().map(|host| {
-    if let Some(port) = uri.port_u16() {
-      format!("{}:{}", host, port)
-    } else {
-      host.to_string()
-    }
-  });
-  if let Some(host) = uri_host {
-    return Ok(host);
-  }
-  // If uri host is not available, use host header
-  host_header_value
-    .map(|h| h.to_str().map(|s| s.to_string()))
-    .transpose()?
-    .ok_or_else(|| anyhow!("No host found in URI or Host header"))
-}
-
-/* --------------------------------------------------------------------------------------------------------- */
-/* Hop-by-hop header handling                                                                              */
-/* --------------------------------------------------------------------------------------------------------- */
-
-/// Remove connection header
-pub(super) fn remove_connection_header(headers: &mut HeaderMap) {
-  if let Some(values) = headers.get(header::CONNECTION) {
-    if let Ok(v) = values.clone().to_str() {
-      let keys = v.split(',').map(|m| m.trim()).filter(|m| !m.is_empty());
-      for m in keys {
-        headers.remove(m);
-      }
-    }
-  }
-}
-
-/// Hop header values which are removed at proxy
-const HOP_HEADERS: &[&str] = &[
-  "connection",
-  "te",
-  "trailer",
-  "keep-alive",
-  "proxy-connection",
-  "proxy-authenticate",
-  "proxy-authorization",
-  "transfer-encoding",
-  "upgrade",
-];
-
-/// Remove hop headers
-pub(super) fn remove_hop_header(headers: &mut HeaderMap) {
-  HOP_HEADERS.iter().for_each(|key| {
-    headers.remove(*key);
-  });
-}
-
-/// Extract upgrade header value if exist
-pub(super) fn extract_upgrade(headers: &HeaderMap) -> Option<String> {
-  if let Some(c) = headers.get(header::CONNECTION) {
-    if c
-      .to_str()
-      .unwrap_or("")
-      .split(',')
-      .any(|w| w.trim().eq_ignore_ascii_case(header::UPGRADE.as_str()))
-    {
-      if let Some(Ok(m)) = headers.get(header::UPGRADE).map(|u| u.to_str()) {
-        debug!("Upgrade in request header: {}", m);
-        return Some(m.to_owned());
-      }
-    }
-  }
-  None
-}
-
-/* --------------------------------------------------------------------------------------------------------- */
 #[cfg(test)]
 mod tests {
   use super::*;
@@ -1036,7 +828,11 @@ mod tests {
     .unwrap();
 
     let normalized_chain = vec![ForwardedEntry {
-      for_node: ForwardedNode::Ip("2001:db8::1".parse().unwrap(), None),
+      for_node: ForwardedNode {
+        ip: Some("2001:db8::1".parse().unwrap()),
+        port: None,
+        raw: ForwardedNodeRaw::Ip,
+      },
       proto: Some("https".into()),
       host: Some("app.example".into()),
       by: None,
