+++ rpxy-lib/src/message_handler/http_log.rs
@@ -1,5 +1,5 @@
 use super::canonical_address::ToCanonical;
-use crate::{log::*, message_handler::utils_headers};
+use crate::{log::*, message_handler::header_ops};
 use http::header;
 use std::net::SocketAddr;
 
@@ -31,7 +31,7 @@ impl<T> From<&http::Request<T>> for HttpMessageLog {
         .to_string()
     };
     let host =
-      utils_headers::host_from_uri_or_host_header(req.uri(), req.headers().get(header::HOST).cloned()).unwrap_or_default();
+      header_ops::host_from_uri_or_host_header(req.uri(), req.headers().get(header::HOST).cloned()).unwrap_or_default();
 
     Self {
       // tls_server_name: "".to_string(),
